Move diffPath entrypoints to diff.ts
