+++ src/diffPath/__tests__/addedDiff.test.ts
@@ -1,4 +1,4 @@
-import { addedDiff } from '../addedDiff';
+import { addedDiff } from '../diff';
 
 describe('addedDiff', () => {
   it('should return safe JSONPaths', () => {
