+++ src/diffPath/__tests__/deletedDiff.test.ts
@@ -1,4 +1,4 @@
-import { deletedDiff } from '../deletedDiff';
+import { deletedDiff } from '../diff';
 
 describe('deletedDiff', () => {
   it('should return safe JSONPaths', () => {
